+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 5.9.1
+
 - [browser] ref: Fix regression with bundle size
 
 ## 5.9.0